refactor: clean up rows
